Fix an int variable to double

+++ src/main/java/pt/upskill/Calculadora.java
@@ -7,7 +7,7 @@ public class Calculadora {
         result = 0;
     }
 
-    public double subtraction(double num1, int num2) {
+    public double subtraction(double num1, double num2) {
         return num1 - num2;
     }
     public double subtraction(double num1) {
